requirements.txt
SimpleITK
numpy
timm
torchvision
openslide-python